+++ src/styles/JournalStats.scss
@@ -1,7 +1,7 @@
 @import 'variables.scss';
 @import 'responsive.scss';
 
-.journal-stats {
+.stats-tab {
   background-color: #1a1a1a;
   padding: 1em;
   margin-bottom: 1em;
